backend/requirements.txt
numpy
tqdm
transformers>=4.19.0
ffmpeg-python==0.2.0
pyaudio
SpeechRecognition
pydub
git+https://github.com/openai/whisper.git
--extra-index-url https://download.pytorch.org/whl/cu118
torch
torchvision
torchaudio
flask
flask_cors
openai
python-dotenv
TTS